Added data handlers. IMDB data extraction heavily refactored
