+++ util/language.py
@@ -3,9 +3,15 @@ try:
 except ImportError:
     print '[!] You need to install spaCy! Visit spacy.io/#install'
 
-
 # Spacy.en provides a faster tokenizer than nltk
 nlp = English()
+
+def parse_paragraph(txt):
+    """
+    Takes a text and returns a list of lists of tokens, where each sublist is a sentence
+    """
+    sentences = nlp(u'' + txt.decode('ascii', errors='ignore')).sents
+    return [[t.text for t in s] for s in sentences]
 
 def tokenize_text(text):
     """
